gait_recognition/requirements.txt
opencv-python-headless==4.8.1.78
numpy>=1.26.0
tensorflow>=2.14.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
matplotlib>=3.7.2
scikit-learn>=1.3.0
tqdm>=4.65.0 